Add news and publication models, serializers, views, and admin configurations
