+++ apps/news/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+
+class NewsConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'apps.news'
