+++ apps/projects/views.py
@@ -1,0 +1,1 @@
+from django.shortcuts import render
