+++ apps/staff/admin.py
@@ -1,0 +1,9 @@
+from django.contrib import admin
+
+from .models import Staff
+
+
+@admin.register(Staff)
+class StaffAdmin(admin.ModelAdmin):
+    list_display = ("name", "position")
+    search_fields = ("name", "position")
